backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic==2.7.2
pydantic-settings==2.2.1
python-multipart==0.0.9
anthropic>=0.45.0
PyJWT==2.8.0
google-genai>=1.0.0
redis>=5.0.0
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pytz
